LayoutTests/http/tests/navigation/https-in-page-cache-expected.txt

ALERT: This page is https and has the no-store cache-control directive. It should NOT go in to the page cache.
ALERT: The page was reloaded on back, not from the page cache. Good job. Running part 2 of the test.
ALERT: This page is https and has the no-cache cache-control directive. It should go in to the page cache.
ALERT: The page was restored from the page cache. Good job!. Running part 3 of the test.
ALERT: This page is https and should go in to the page cache.
ALERT: The page was restored from the page cache. Good job!
